refactor: User 클래스 교체
