+++ api/src/main/java/org/yapp/domain/user/dao/UserRepository.java
@@ -1,9 +1,10 @@
 package org.yapp.domain.user.dao;
 
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.yapp.domain.user.domain.User;
+import org.yapp.domain.user.User;
+
+import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByOauthId(String oauthId);
+  Optional<User> findByOauthId(String oauthId);
 }
