+++ common/src/main/java/org/yapp/domain/user/User.java
@@ -1,0 +1,44 @@
+package org.yapp.domain.user;
+
+import org.yapp.domain.profile.Profile;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Table(name = "user_common")
+@Entity
+@Getter
+@NoArgsConstructor
+public class User {
+  @Id
+  @Column(name = "user_id")
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @Column(name = "oauth_id")
+  private String oauthId;
+
+  @Column(name = "name")
+  private String name;
+
+  @OneToOne
+  @JoinColumn(name = "profile_id", unique = true)  // User가 profile_id를 FK로 가짐
+  private Profile profile;
+
+  @Builder
+  public User(String oauthId, String name, Profile profile) {
+    this.oauthId = oauthId;
+    this.name = name;
+    this.profile = profile;
+  }
+}
+
